requirements.txt
python-dotenv>=0.19.0
openai>=1.0.0
anthropic>=0.3.0
google-generativeai>=0.3.0
deepseek-ai>=0.0.1
streamlit>=1.32.0
watchdog>=3.0.0
pymongo>=4.6.0